+++ src/main/java/es/us/isa/idlreasonerchoco/analyzer/Analyzer.java
@@ -44,8 +44,7 @@ public abstract class Analyzer {
     
     public abstract void updateData(Map<String, List<String>> data) throws IDLException;
 
-    public abstract Map<String, Map<String, List<String>>> getRequestExplanation(Map<String, String> request) throws IDLException;
+    public abstract Map<String, Map<String, List<String>>> getExplanation(Map<String, String> request) throws IDLException;
 
-    public abstract Map<String, Map<String, List<String>>> getOASOperationExplanation() throws IDLException;
 
 }
