+++ src/test/java/es/us/isa/idlreasonerchoco/explanation/RequestExplanationTest.java
@@ -1,4 +1,4 @@
-package es.us.isa.idlreasonerchoco;
+package es.us.isa.idlreasonerchoco.explanation;
 
 import es.us.isa.idlreasonerchoco.analyzer.Analyzer;
 import es.us.isa.idlreasonerchoco.analyzer.OASAnalyzer;
@@ -37,7 +37,7 @@ public class RequestExplanationTest {
         request.put("p1", "not boolean");
         assertFalse(analyzer.isValidRequest(request), "The request should be NOT valid");
 
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
         assertNotNull(result.get("Error"), "The Invalid Request Params Conflicts should not be empty");
 
@@ -52,10 +52,10 @@ public class RequestExplanationTest {
         request.put("p1", "a string");
         assertTrue(analyzer.isValidRequest(request), "The request should be VALID");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertTrue(result.get("InvalidRequestParams").isEmpty(), "The Invalid Request Params Conflicts should not be empty");
+        assertTrue(result.get("IDLConflicts").isEmpty(), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: one_param_string_no_deps_valid.");
     }
@@ -66,10 +66,10 @@ public class RequestExplanationTest {
         Map<String, String> request = new HashMap<>();
         assertFalse(analyzer.isValidRequest(request), "The request should be NOT valid");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNotNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-       // assertNotNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertNotNull(result.get("InvalidRequestParams"), "The Invalid Request Params Conflicts should not be empty");
+       // assertNotNull(result.get("IDLConflicts"), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: one_param_string_no_deps_invalid.");
     }
@@ -81,10 +81,10 @@ public class RequestExplanationTest {
         request.put("p1", "10");
         assertTrue(analyzer.isValidRequest(request), "The request should be VALID");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertTrue(result.get("InvalidRequestParams").isEmpty(), "The Invalid Request Params Conflicts should not be empty");
+        assertTrue(result.get("IDLConflicts").isEmpty(), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: one_param_int_no_deps_valid.");
     }
@@ -97,10 +97,10 @@ public class RequestExplanationTest {
         // request.put("p1", "not an integer");
        // assertFalse(analyzer.isValidRequest(request), "The request should be NOT valid");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-       // assertNotNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-       // assertNotNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+       // assertNotNull(result.get("InvalidRequestParams"), "The Invalid Request Params Conflicts should not be empty");
+       // assertNotNull(result.get("IDLConflicts"), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: one_param_int_no_deps_invalid.");
     }
@@ -112,10 +112,10 @@ public class RequestExplanationTest {
         request.put("p1", "value1");
         assertTrue(analyzer.isValidRequest(request), "The request should be VALID");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertTrue(result.get("InvalidRequestParams").isEmpty(), "The Invalid Request Params Conflicts should not be empty");
+        assertTrue(result.get("IDLConflicts").isEmpty(), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: one_param_enum_string_no_deps_valid.");
     }
@@ -127,10 +127,10 @@ public class RequestExplanationTest {
         request.put("p1", "string not in enum alternatives");
         assertFalse(analyzer.isValidRequest(request), "The request should be NOT valid");
        
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNotNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-       // assertNotNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertNotNull(result.get("InvalidRequestParams"), "The Invalid Request Params Conflicts should not be empty");
+       // assertNotNull(result.get("IDLConflicts"), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: one_param_enum_string_no_deps_invalid.");
     }
@@ -142,10 +142,10 @@ public class RequestExplanationTest {
         request.put("p1", "1");
         assertTrue(analyzer.isValidRequest(request), "The request should be VALID");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertTrue(result.get("InvalidRequestParams").isEmpty(), "The Invalid Request Params Conflicts should not be empty");
+        assertTrue(result.get("IDLConflicts").isEmpty(), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: one_param_enum_int_no_deps_valid.");
     }
@@ -157,10 +157,10 @@ public class RequestExplanationTest {
         request.put("p1", "6");
         assertFalse(analyzer.isValidRequest(request), "The request should be NOT valid");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNotNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-       // assertNotNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertNotNull(result.get("InvalidRequestParams"), "The Invalid Request Params Conflicts should not be empty");
+       // assertNotNull(result.get("IDLConflicts"), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: one_param_enum_int_no_deps_invalid.");
     }
@@ -173,11 +173,11 @@ public class RequestExplanationTest {
        // request.put("p2", "a string");
         assertFalse(analyzer.isValidRequest(request), "The request should be VALID");
        
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
         
-        assertNotNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNotNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertNotNull(result.get("InvalidRequestParams"), "The Invalid Request Params Conflicts should not be empty");
+        assertNotNull(result.get("IDLConflicts"), "The IDL Conflicts should not be empty");
 
         
         System.out.println("Test passed: one_dep_requires_valid.");
@@ -193,10 +193,10 @@ public class RequestExplanationTest {
         request.put("p5", "2");
         assertFalse(analyzer.isValidRequest(request), "The request should be NOT valid");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        //assertNotNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        //assertNotNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        //assertNotNull(result.get("InvalidRequestParams"), "The Invalid Request Params Conflicts should not be empty");
+        //assertNotNull(result.get("IDLConflicts"), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: one_dep_requires_invalid.");
     }
@@ -208,10 +208,10 @@ public class RequestExplanationTest {
         request.put("p1", "true");
         assertTrue(analyzer.isValidRequest(request), "The request should be VALID");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertTrue(result.get("InvalidRequestParams").isEmpty(), "The Invalid Request Params Conflicts should not be empty");
+        assertTrue(result.get("IDLConflicts").isEmpty(), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: one_dep_or_valid.");
     }
@@ -222,10 +222,10 @@ public class RequestExplanationTest {
         Map<String, String> request = new HashMap<>();
         assertFalse(analyzer.isValidRequest(request), "The request should be NOT valid");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-      //  assertNotNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-     //   assertNotNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+      //  assertNotNull(result.get("InvalidRequestParams"), "The Invalid Request Params Conflicts should not be empty");
+     //   assertNotNull(result.get("IDLConflicts"), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: one_dep_or_invalid.");
     }
@@ -238,10 +238,10 @@ public class RequestExplanationTest {
         request.put("p3", "100");
         assertTrue(analyzer.isValidRequest(request), "The request should be VALID");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertTrue(result.get("InvalidRequestParams").isEmpty(), "The Invalid Request Params Conflicts should not be empty");
+        assertTrue(result.get("IDLConflicts").isEmpty(), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: one_dep_onlyone_valid.");
     }
@@ -254,10 +254,10 @@ public class RequestExplanationTest {
         request.put("p2", "string");
         assertFalse(analyzer.isValidRequest(request), "The request should be NOT valid");
        
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNotNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNotNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertNotNull(result.get("InvalidRequestParams"), "The Invalid Request Params Conflicts should not be empty");
+        assertNotNull(result.get("IDLConflicts"), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: one_dep_onlyone_invalid.");
     }
@@ -271,10 +271,10 @@ public class RequestExplanationTest {
         request.put("p5", "1");
         assertTrue(analyzer.isValidRequest(request), "The request should be VALID");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertTrue(result.get("InvalidRequestParams").isEmpty(), "The Invalid Request Params Conflicts should not be empty");
+        assertTrue(result.get("IDLConflicts").isEmpty(), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: one_dep_allornone_valid.");
     }
@@ -289,7 +289,7 @@ public class RequestExplanationTest {
         request.put("p5", "1");
         assertFalse(analyzer.isValidRequest(request), "The request should be NOT valid");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
         System.out.println("Test passed: one_dep_allornone_invalid.");
         
@@ -301,10 +301,10 @@ public class RequestExplanationTest {
         Map<String, String> request = new HashMap<>();
         assertTrue(analyzer.isValidRequest(request), "The request should be VALID");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertTrue(result.get("InvalidRequestParams").isEmpty(), "The Invalid Request Params Conflicts should not be empty");
+        assertTrue(result.get("IDLConflicts").isEmpty(), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: one_dep_zeroorone_valid.");
     }
@@ -320,10 +320,10 @@ public class RequestExplanationTest {
         request.put("p5", "1");
         assertFalse(analyzer.isValidRequest(request), "The request should be NOT valid");
        
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNotNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNotNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertNotNull(result.get("InvalidRequestParams"), "The Invalid Request Params Conflicts should not be empty");
+        assertNotNull(result.get("IDLConflicts"), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: one_dep_zeroorone_invalid.");
     }
@@ -336,10 +336,10 @@ public class RequestExplanationTest {
         request.put("p5", "1");
         assertTrue(analyzer.isValidRequest(request), "The request should be VALID");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertTrue(result.get("InvalidRequestParams").isEmpty(), "The Invalid Request Params Conflicts should not be empty");
+        assertTrue(result.get("IDLConflicts").isEmpty(), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: one_dep_arithrel_valid.");
     }
@@ -352,10 +352,10 @@ public class RequestExplanationTest {
         request.put("p5", "1");
         assertFalse(analyzer.isValidRequest(request), "The request should be NOT valid");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNotNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNotNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertNotNull(result.get("InvalidRequestParams"), "The Invalid Request Params Conflicts should not be empty");
+        assertNotNull(result.get("IDLConflicts"), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: one_dep_arithrel_invalid.");
     }
@@ -370,10 +370,10 @@ public class RequestExplanationTest {
         request.put("p5", "4");
         assertTrue(analyzer.isValidRequest(request), "The request should be VALID");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should be empty");
+        assertTrue(result.get("InvalidRequestParams").isEmpty(), "The Invalid Request Params Conflicts should not be empty");
+        assertTrue(result.get("IDLConflicts").isEmpty(), "The IDL Conflicts should be empty");
         
         System.out.println("Test passed: one_dep_complex_valid.");
     }
@@ -387,10 +387,10 @@ public class RequestExplanationTest {
         request.put("p3", "-1000");
         assertFalse(analyzer.isValidRequest(request), "The request should be NOT valid");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNotNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNotNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertNotNull(result.get("InvalidRequestParams"), "The Invalid Request Params Conflicts should not be empty");
+        assertNotNull(result.get("IDLConflicts"), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: one_dep_complex_invalid.");
     }
@@ -406,10 +406,10 @@ public class RequestExplanationTest {
         request.put("p5", "value1");
         assertTrue(analyzer.isValidRequest(request), "The request should be VALID");
        
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertTrue(result.get("InvalidRequestParams").isEmpty(), "The Invalid Request Params Conflicts should not be empty");
+        assertTrue(result.get("IDLConflicts").isEmpty(), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: combinatorial1_valid.");
     }
@@ -425,10 +425,10 @@ public class RequestExplanationTest {
         request.put("p5", "value5"); // Violates this dependency: p1==p5;
         assertFalse(analyzer.isValidRequest(request), "The request should be NOT valid");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNotNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNotNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertNotNull(result.get("InvalidRequestParams"), "The Invalid Request Params Conflicts should not be empty");
+        assertNotNull(result.get("IDLConflicts"), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: combinatorial1_invalid.");
         
@@ -453,11 +453,11 @@ public class RequestExplanationTest {
         request.put("p3", "value1");
         assertTrue(analyzer.isValidRequest(request), "The request should be VALID");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
 
-        assertNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should  be empty");
-        assertNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should be empty");
+        assertTrue(result.get("InvalidRequestParams").isEmpty(), "The Invalid Request Params Conflicts should  be empty");
+        assertTrue(result.get("IDLConflicts").isEmpty(), "The IDL Conflicts should be empty");
 
         System.out.println("Test passed: combinatorial3_Valid.");
     }
@@ -473,10 +473,10 @@ public class RequestExplanationTest {
         request.put("p5", "3"); // Violates this dependency: p4>=p5;
         assertFalse(analyzer.isValidRequest(request), "The request should be NOT valid");
        
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNotNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNotNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertNotNull(result.get("InvalidRequestParams"), "The Invalid Request Params Conflicts should not be empty");
+        assertNotNull(result.get("IDLConflicts"), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: combinatorial3_invalid.");
            
@@ -490,10 +490,10 @@ public class RequestExplanationTest {
         request.put("p3", "100000");
         assertTrue(analyzer.isValidRequest(request), "The request should be VALID");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertTrue(result.get("InvalidRequestParams").isEmpty(), "The Invalid Request Params Conflicts should not be empty");
+        assertTrue(result.get("IDLConflicts").isEmpty(), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: combinatorial4_valid.");
     }
@@ -505,10 +505,10 @@ public class RequestExplanationTest {
          
         assertFalse(analyzer.isValidRequest(request), "The request should be NOT valid");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNotNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNotNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertNotNull(result.get("InvalidRequestParams"), "The Invalid Request Params Conflicts should not be empty");
+        assertNotNull(result.get("IDLConflicts"), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: combinatorial4_invalid.");
   
@@ -530,10 +530,10 @@ public class RequestExplanationTest {
         request.put("p10", "value2");
         assertTrue(analyzer.isValidRequest(request), "The request should be VALID");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertTrue(result.get("InvalidRequestParams").isEmpty(), "The Invalid Request Params Conflicts should not be empty");
+        assertTrue(result.get("IDLConflicts").isEmpty(), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: combinatorial5_valid.");
     }
@@ -555,10 +555,10 @@ public class RequestExplanationTest {
         request.put("p10", "value2");
         assertFalse(analyzer.isValidRequest(request), "The request should be NOT valid");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNotNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNotNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertNotNull(result.get("InvalidRequestParams"), "The Invalid Request Params Conflicts should not be empty");
+        assertNotNull(result.get("IDLConflicts"), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: combinatorial5_invalid.");
  
@@ -578,10 +578,10 @@ public class RequestExplanationTest {
         request.put("p10", "something");
         assertTrue(analyzer.isValidRequest(request), "The request should be VALID");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertTrue(result.get("InvalidRequestParams").isEmpty(), "The Invalid Request Params Conflicts should not be empty");
+        assertTrue(result.get("IDLConflicts").isEmpty(), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: combinatorial8_valid.");
     }
@@ -600,10 +600,10 @@ public class RequestExplanationTest {
         request.put("p10", "something different from p8"); // Violates this dependency: AllOrNone(p6!=p8, p8==p10);
         assertFalse(analyzer.isValidRequest(request), "The request should be NOT valid");
         
-        Map<String, Map<String, List<String>>> result =  analyzer.getRequestExplanation(request);
+        Map<String, Map<String, List<String>>> result =  analyzer.getExplanation(request);
         System.out.println("result: " + result);
-        assertNotNull(result.get("InvalidRequestParamsExplanation"), "The Invalid Request Params Conflicts should not be empty");
-        assertNotNull(result.get("IDLConflictsExplanation"), "The IDL Conflicts should not be empty");
+        assertNotNull(result.get("InvalidRequestParams"), "The Invalid Request Params Conflicts should not be empty");
+        assertNotNull(result.get("IDLConflicts"), "The IDL Conflicts should not be empty");
         
         System.out.println("Test passed: combinatorial8_invalid.");
     }
